Moved to vite and fixed some bugs
The create-react-app (webpack) is old and slow so I moved to vite which faster and simpler.
Updated dependencies and removed unused ones.
Fixed the '0' appearing on the search boxes when they are empty or there are 0 results.
Fixed the side navbar top being cut off.
Made the credits card use the drop-shadow filter so when the pic isn't a full square it casts correctly around the pic.
Improved the layout style of release accordions markdown.
Moved the pagination (used in Releases) to its on file.